Small syntax changes
Map pairs are now coupled by '=' instead of '=>' to reserve arrow glyphs exclusively for messaging usages as opposed to this mapping usage, despite its conventionality. The '=' is also consistent with the connotation of assignment within a data structure,  and with the assignment statement that can re-map keys.

Scan is now coupled to its consumer by a new glyph '>>' to indicate plurality and differentiate from the the reply-handling usage, which is semantically very different.